+++ scenario-compiler/src/main/scala/pl/touk/nussknacker/engine/definition/model/ModelDefinition.scala
@@ -3,7 +3,6 @@ package pl.touk.nussknacker.engine.definition.model
 import pl.touk.nussknacker.engine.api.component.ComponentId
 import pl.touk.nussknacker.engine.api.component.ComponentType.ComponentType
 import pl.touk.nussknacker.engine.api.process.ClassExtractionSettings
-import pl.touk.nussknacker.engine.definition.component.Components.ComponentDefinitionExtractionMode
 import pl.touk.nussknacker.engine.definition.component.{ComponentDefinitionWithImplementation, Components}
 import pl.touk.nussknacker.engine.definition.globalvariables.ExpressionConfigDefinition
 
@@ -13,21 +12,17 @@ case class ModelDefinition private (
     settings: ClassExtractionSettings
 ) {
 
-  import pl.touk.nussknacker.engine.util.Implicits._
-
   def withComponent(component: ComponentDefinitionWithImplementation): ModelDefinition = {
     withComponents(List(component))
   }
 
   def withComponents(componentsToAdd: Components): ModelDefinition = {
     val newComponents = components.withComponents(componentsToAdd)
-    checkDuplicates(newComponents)
     copy(components = newComponents)
   }
 
   def withComponents(componentsToAdd: List[ComponentDefinitionWithImplementation]): ModelDefinition = {
     val newComponents = components.withComponents(componentsToAdd)
-    checkDuplicates(newComponents)
     copy(components = newComponents)
   }
 
@@ -40,28 +35,6 @@ case class ModelDefinition private (
 
   def filterComponents(predicate: ComponentDefinitionWithImplementation => Boolean): ModelDefinition =
     copy(components = components.filter(predicate))
-
-  private def checkDuplicates(components: Components): Unit = {
-
-    def findDuplicates(values: List[ComponentDefinitionWithImplementation]) = {
-      values
-        .map(component => component.id -> component)
-        .toGroupedMap
-        .filter(_._2.size > 1)
-        .keys
-        .toList
-        .sorted
-    }
-
-    val componentsDuplicates      = findDuplicates(components.components)
-    val basicComponentsDuplicates = components.basicComponents.map(findDuplicates).getOrElse(List.empty)
-
-    if (componentsDuplicates.nonEmpty) {
-      throw new DuplicatedComponentsException(componentsDuplicates)
-    } else if (basicComponentsDuplicates.nonEmpty) {
-      throw new DuplicatedComponentsException(basicComponentsDuplicates)
-    }
-  }
 
 }
 
@@ -76,14 +49,6 @@ object ModelDefinition {
       components: Components,
       expressionConfig: ExpressionConfigDefinition,
       settings: ClassExtractionSettings,
-  ): ModelDefinition = {
-    val componentDefinitionExtractionMode = components.basicComponents match {
-      case Some(_) => ComponentDefinitionExtractionMode.FinalAndBasicDefinitions
-      case None    => ComponentDefinitionExtractionMode.FinalDefinition
-    }
-
-    new ModelDefinition(Components.empty(componentDefinitionExtractionMode), expressionConfig, settings)
-      .withComponents(components)
-  }
+  ): ModelDefinition = new ModelDefinition(components, expressionConfig, settings)
 
 }